+++ express-zod-api/tests/documentation-helpers.spec.ts
@@ -397,6 +397,7 @@ describe("Documentation helpers", () => {
       z.record(z.literal("testing"), z.boolean()),
       z.record(z.literal("one").or(z.literal("two")), z.boolean()),
       z.record(z.any()), // Issue #900
+      z.record(z.string().regex(/x-\w+/), z.boolean()),
     ])(
       "should set properties+required or additionalProperties props %#",
       (schema) => {
@@ -408,6 +409,16 @@ describe("Documentation helpers", () => {
   describe("depictArray()", () => {
     test("should set type:array and pass items depiction", () => {
       expect(depictArray(z.array(z.boolean()), requestCtx)).toMatchSnapshot();
+    });
+
+    test.each([
+      z.boolean().array().min(3),
+      z.boolean().array().max(5),
+      z.boolean().array().min(3).max(5),
+      z.boolean().array().length(4),
+      z.array(z.boolean()).nonempty(),
+    ])("should reflect min/max/exact length of the array %#", (schema) => {
+      expect(depictArray(schema, requestCtx)).toMatchSnapshot();
     });
   });